src: introduce node_bootstrap and node_process
This continues the reorganization of node.cc by splitting the
implementations of most process object functions and properties
out to a separate node_process.cc and creates a separate
temporary bootstrapper object used during Node.js bootstrap
to avoid the need for deleting/overwriting process properties
set during bootstrap.
