+++ lib/internal/bootstrap/node.js
@@ -10,10 +10,16 @@
 // process.binding(), process._linkedBinding(), internalBinding() and
 // NativeModule. And then { internalBinding, NativeModule } will be passed
 // into this bootstrapper to bootstrap Node.js core.
-
+//
+// The bootstrapper argument is an object containing functions and properties
+// that are only relevant during the bootstrap process. A permanent reference
+// to this object must not be maintained, allowing the object to gc when it is
+// no longer needed.
 'use strict';
 
-(function bootstrapNodeJSCore(process, { internalBinding, NativeModule }) {
+(function bootstrapNodeJSCore(process,
+                              bootstrapper,
+                              { internalBinding, NativeModule }) {
   const exceptionHandlerState = { captureFn: null };
 
   function startup() {
@@ -36,11 +42,11 @@
     const _process = NativeModule.require('internal/process');
     _process.setupConfig(NativeModule._source);
     _process.setupSignalHandlers();
-    _process.setupUncaughtExceptionCapture(exceptionHandlerState);
+    _process.setupUncaughtExceptionCapture(exceptionHandlerState, bootstrapper);
     NativeModule.require('internal/process/warning').setup();
-    NativeModule.require('internal/process/next_tick').setup();
+    NativeModule.require('internal/process/next_tick').setup(bootstrapper);
     NativeModule.require('internal/process/stdio').setup();
-    NativeModule.require('internal/process/methods').setup();
+    NativeModule.require('internal/process/methods').setup(bootstrapper);
 
     const perf = process.binding('performance');
     const {
@@ -55,9 +61,9 @@
       NODE_PERFORMANCE_MILESTONE_PRELOAD_MODULE_LOAD_END
     } = perf.constants;
 
-    _process.setup_hrtime();
-    _process.setup_cpuUsage();
-    _process.setupMemoryUsage();
+    _process.setup_hrtime(bootstrapper);
+    _process.setup_cpuUsage(bootstrapper);
+    _process.setupMemoryUsage(bootstrapper);
     _process.setupKillAndExit();
     if (global.__coverage__)
       NativeModule.require('internal/process/write-coverage').setup();
@@ -79,7 +85,7 @@
     }
 
     _process.setupChannel();
-    _process.setupRawDebug();
+    _process.setupRawDebug(bootstrapper);
 
     const browserGlobals = !process._noBrowserGlobals;
     if (browserGlobals) {
@@ -116,9 +122,14 @@
         'DeprecationWarning', 'DEP0062', startup, true);
     }
 
-    if (process.binding('config').experimentalModules ||
-        process.binding('config').experimentalVMModules) {
-      if (process.binding('config').experimentalModules) {
+    const {
+      experimentalModules,
+      experimentalVMModules,
+      pendingDeprecation
+    } = process.binding('config');
+
+    if (experimentalModules || experimentalVMModules) {
+      if (experimentalModules) {
         process.emitWarning(
           'The ESM module loader is experimental.',
           'ExperimentalWarning', undefined);
@@ -129,7 +140,6 @@
     {
       // Install legacy getters on the `util` binding for typechecking.
       // TODO(addaleax): Turn into a full runtime deprecation.
-      const { pendingDeprecation } = process.binding('config');
       const { deprecate } = NativeModule.require('internal/util');
       const utilBinding = process.binding('util');
       const types = internalBinding('types');
@@ -294,7 +304,7 @@
   }
 
   function setupProcessObject() {
-    process._setupProcessObject(pushValueToArray);
+    bootstrapper._setupProcessObject(pushValueToArray);
 
     function pushValueToArray() {
       for (var i = 0; i < arguments.length; i++)
