+++ lib/internal/process.js
@@ -25,10 +25,7 @@ process.assert = deprecate(
   'DEP0100');
 
 // Set up the process.cpuUsage() function.
-function setup_cpuUsage() {
-  // Get the native function, which will be replaced with a JS version.
-  const _cpuUsage = process.cpuUsage;
-
+function setup_cpuUsage({ _cpuUsage }) {
   // Create the argument array that will be passed to the native function.
   const cpuValues = new Float64Array(2);
 
@@ -92,8 +89,7 @@ function setup_cpuUsage() {
 // The 3 entries filled in by the original process.hrtime contains
 // the upper/lower 32 bits of the second part of the value,
 // and the remaining nanoseconds of the value.
-function setup_hrtime() {
-  const _hrtime = process.hrtime;
+function setup_hrtime({ _hrtime }) {
   const hrValues = new Uint32Array(3);
 
   process.hrtime = function hrtime(time) {
@@ -120,12 +116,11 @@ function setup_hrtime() {
   };
 }
 
-function setupMemoryUsage() {
-  const memoryUsage_ = process.memoryUsage;
+function setupMemoryUsage({ _memoryUsage }) {
   const memValues = new Float64Array(4);
 
   process.memoryUsage = function memoryUsage() {
-    memoryUsage_(memValues);
+    _memoryUsage(memValues);
     return {
       rss: memValues[0],
       heapTotal: memValues[1],
@@ -245,23 +240,21 @@ function setupChannel() {
 }
 
 
-function setupRawDebug() {
-  const rawDebug = process._rawDebug;
+function setupRawDebug({ _rawDebug }) {
   process._rawDebug = function() {
-    rawDebug(util.format.apply(null, arguments));
+    _rawDebug(util.format.apply(null, arguments));
   };
 }
 
 
-function setupUncaughtExceptionCapture(exceptionHandlerState) {
-  // This is a typed array for faster communication with JS.
-  const shouldAbortOnUncaughtToggle = process._shouldAbortOnUncaughtToggle;
-  delete process._shouldAbortOnUncaughtToggle;
-
+function setupUncaughtExceptionCapture(exceptionHandlerState,
+                                       { _shouldAbortOnUncaughtToggle }) {
+  // _shouldAbortOnUncaughtToggleis a typed array for faster
+  // communication with JS.
   process.setUncaughtExceptionCaptureCallback = function(fn) {
     if (fn === null) {
       exceptionHandlerState.captureFn = fn;
-      shouldAbortOnUncaughtToggle[0] = 1;
+      _shouldAbortOnUncaughtToggle[0] = 1;
       return;
     }
     if (typeof fn !== 'function') {
@@ -271,7 +264,7 @@ function setupUncaughtExceptionCapture(exceptionHandlerState) {
       throw new ERR_UNCAUGHT_EXCEPTION_CAPTURE_ALREADY_SET();
     }
     exceptionHandlerState.captureFn = fn;
-    shouldAbortOnUncaughtToggle[0] = 0;
+    _shouldAbortOnUncaughtToggle[0] = 0;
   };
 
   process.hasUncaughtExceptionCaptureCallback = function() {
