+++ lib/internal/process/methods.js
@@ -9,55 +9,40 @@ const {
   validateUint32
 } = require('internal/validators');
 
-function setupProcessMethods() {
+function setupProcessMethods(bootstrapper) {
   // Non-POSIX platforms like Windows don't have certain methods.
-  if (process.setgid !== undefined) {
-    setupPosixMethods();
+  if (bootstrapper._setgid !== undefined) {
+    setupPosixMethods(bootstrapper);
   }
 
-  const {
-    chdir: _chdir,
-    umask: _umask,
-  } = process;
+  setupChdir(bootstrapper);
+  setupUmask(bootstrapper);
+}
 
-  process.chdir = chdir;
-  process.umask = umask;
-
-  function chdir(directory) {
+function setupChdir({ _chdir }) {
+  process.chdir = function chdir(directory) {
     if (typeof directory !== 'string') {
       throw new ERR_INVALID_ARG_TYPE('directory', 'string', directory);
     }
     return _chdir(directory);
-  }
+  };
+}
 
-  function umask(mask) {
+function setupUmask({ _umask }) {
+  process.umask = function umask(mask) {
     if (mask === undefined) {
       // Get the mask
       return _umask(mask);
     }
     mask = validateAndMaskMode(mask, 'mask');
     return _umask(mask);
-  }
+  };
 }
 
-function setupPosixMethods() {
-  const {
-    initgroups: _initgroups,
-    setegid: _setegid,
-    seteuid: _seteuid,
-    setgid: _setgid,
-    setuid: _setuid,
-    setgroups: _setgroups
-  } = process;
+function setupPosixMethods({ _initgroups, _setegid, _seteuid,
+                             _setgid, _setuid, _setgroups }) {
 
-  process.initgroups = initgroups;
-  process.setegid = setegid;
-  process.seteuid = seteuid;
-  process.setgid = setgid;
-  process.setuid = setuid;
-  process.setgroups = setgroups;
-
-  function initgroups(user, extraGroup) {
+  process.initgroups = function initgroups(user, extraGroup) {
     validateId(user, 'user');
     validateId(extraGroup, 'extraGroup');
     // Result is 0 on success, 1 if user is unknown, 2 if group is unknown.
@@ -67,25 +52,25 @@ function setupPosixMethods() {
     } else if (result === 2) {
       throw new ERR_UNKNOWN_CREDENTIAL('Group', extraGroup);
     }
-  }
+  };
 
-  function setegid(id) {
+  process.setegid = function setegid(id) {
     return execId(id, 'Group', _setegid);
-  }
+  };
 
-  function seteuid(id) {
+  process.seteuid = function seteuid(id) {
     return execId(id, 'User', _seteuid);
-  }
+  };
 
-  function setgid(id) {
+  process.setgid = function setgid(id) {
     return execId(id, 'Group', _setgid);
-  }
+  };
 
-  function setuid(id) {
+  process.setuid = function setuid(id) {
     return execId(id, 'User', _setuid);
-  }
+  };
 
-  function setgroups(groups) {
+  process.setgroups = function setgroups(groups) {
     if (!Array.isArray(groups)) {
       throw new ERR_INVALID_ARG_TYPE('groups', 'Array', groups);
     }
@@ -98,7 +83,7 @@ function setupPosixMethods() {
     if (result > 0) {
       throw new ERR_UNKNOWN_CREDENTIAL('Group', groups[result - 1]);
     }
-  }
+  };
 
   function execId(id, type, method) {
     validateId(id, 'id');
