+++ lib/internal/process/next_tick.js
@@ -2,7 +2,7 @@
 
 exports.setup = setupNextTick;
 
-function setupNextTick() {
+function setupNextTick({ _setupNextTick, _setupPromises }) {
   const {
     getDefaultTriggerAsyncId,
     newAsyncId,
@@ -14,10 +14,10 @@ function setupNextTick() {
     emitDestroy,
     symbols: { async_id_symbol, trigger_async_id_symbol }
   } = require('internal/async_hooks');
-  const promises = require('internal/process/promises');
+  const emitPromiseRejectionWarnings =
+    require('internal/process/promises').setup(_setupPromises);
   const { ERR_INVALID_CALLBACK } = require('internal/errors').codes;
   const FixedQueue = require('internal/fixed_queue');
-  const { emitPromiseRejectionWarnings } = promises;
 
   // tickInfo is used so that the C++ code in src/node.cc can
   // have easy access to our nextTick state, and avoid unnecessary
@@ -26,7 +26,7 @@ function setupNextTick() {
   const [
     tickInfo,
     runMicrotasks
-  ] = process._setupNextTick(_tickCallback);
+  ] = _setupNextTick(_tickCallback);
 
   // *Must* match Environment::TickInfo::Fields in src/env.h.
   const kHasScheduled = 0;
