+++ lib/internal/process/promises.js
@@ -7,11 +7,12 @@ const pendingUnhandledRejections = [];
 const asyncHandledRejections = [];
 let lastPromiseId = 0;
 
-module.exports = {
-  emitPromiseRejectionWarnings
-};
+exports.setup = setupPromises;
 
-process._setupPromises(unhandledRejection, handledRejection);
+function setupPromises(_setupPromises) {
+  _setupPromises(unhandledRejection, handledRejection);
+  return emitPromiseRejectionWarnings;
+}
 
 function unhandledRejection(promise, reason) {
   maybeUnhandledPromises.set(promise, {
