+++ node.gyp
@@ -322,6 +322,7 @@
         'src/node_api.cc',
         'src/node_api.h',
         'src/node_api_types.h',
+        'src/node_bootstrap.cc',
         'src/node_buffer.cc',
         'src/node_config.cc',
         'src/node_constants.cc',
@@ -336,6 +337,7 @@
         'src/node_platform.cc',
         'src/node_perf.cc',
         'src/node_postmortem_metadata.cc',
+        'src/node_process.cc',
         'src/node_serdes.cc',
         'src/node_trace_events.cc',
         'src/node_types.cc',
@@ -395,6 +397,7 @@
         'src/node_perf_common.h',
         'src/node_persistent.h',
         'src/node_platform.h',
+        'src/node_process.h',
         'src/node_root_certs.h',
         'src/node_version.h',
         'src/node_watchdog.h',
